Fix some UBSAN issues:   *) common.c:2041:13: runtime error: left shift of      1 by 31 places cannot be represented in type 'long'   *) inet_util.c:847:36: runtime error: left shift of negative      value -1

+++ src/inet_util.c
@@ -843,7 +843,7 @@ quit:
  */
 void INET_util_get_mask4 (struct in_addr *out, int bits)
 {
-  *(DWORD*)out = bits ? swap32 (~0 << (32 - bits)) : 0;
+  *(DWORD*)out = bits ? swap32 (~0UL << (DWORD)(32 - bits)) : 0;
 }
 
 /**
